Добавлен контейнер Go приложения и изменена структура проекта
https://github.com/stas12312/my-tickets-bot/issues/252
