+++ my_tickets_bot/bot/src/bot/handlers/poster/show.py
