+++ my_tickets_bot/bot/src/bot/messages/poster.py
