+++ my_tickets_bot/bot/src/bot/paginators/event.py
